backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
scikit-learn
xgboost
pandas
numpy
transformers
torch
pyzbar
opencv-python-headless
python-dotenv
requests
joblib
matplotlib
seaborn
python-jose[cryptography]
passlib[bcrypt]
qrcode
pillow
python-magic
tldextract
lightgbm
